Transfert more power from ToolBox to WpsClient.
