+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/WpsService.java
@@ -1,22 +1,227 @@
 package org.orbisgis.orbistoolbox;
 
+import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.io.IOUtils;
+import org.h2gis.h2spatialapi.DriverFunction;
+import org.h2gis.h2spatialapi.EmptyProgressVisitor;
+import org.h2gis.utilities.JDBCUtilities;
+import org.orbisgis.corejdbc.DataManager;
+import org.orbisgis.dbjobs.api.DriverFunctionContainer;
+import org.orbisgis.frameworkapi.CoreWorkspace;
+import org.orbisgis.orbistoolbox.controller.execution.DataProcessingManager;
 import org.orbisgis.orbistoolbox.model.Process;
 import org.orbisgis.orbistoolbox.controller.process.ProcessIdentifier;
 import org.orbisgis.orbistoolbox.controller.process.ProcessManager;
+import org.orbisgis.orbistoolbox.view.ToolBox;
+import org.osgi.framework.FrameworkUtil;
+import org.osgi.service.component.annotations.Reference;
+import org.slf4j.LoggerFactory;
 
-import java.io.File;
+import java.io.*;
 import java.net.URI;
-import java.util.Map;
+import java.net.URL;
+import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.*;
 
 public class WpsService {
     /** String of the Groovy file extension. */
     public static final String GROOVY_EXTENSION = "groovy";
-
+    private static final String WPS_SCRIPT_FOLDER = "Scripts";
+    private CoreWorkspace coreWorkspace;
+    /** List of process ended, waiting the 5 seconds before being removed.*/
+    boolean multiThreaded = true;
+    /** String reference of the ToolBox used for DockingFrame. */
+    public static final String TOOLBOX_REFERENCE = "orbistoolbox";
+    private static final String TOOLBOX_PROPERTIES = "toolbox.properties";
+    private static final String PROPERTY_SOURCES = "PROPERTY_SOURCES";
+    /** ToolBox properties */
+    private Properties tbProperties;
+    /** True if the database is H2, false otherwise. */
+    private boolean isH2;
+    /** OrbisGIS DataManager. */
+    private static DataManager dataManager;
+    private DataProcessingManager dataProcessingManager;
+    /** OrbisGIS DriverFunctionContainer. */
+    private static DriverFunctionContainer driverFunctionContainer;
     /** Process manager which contains all the loaded scripts. */
     private ProcessManager processManager;
+    private static boolean areScriptsCopied = false;
 
-    public WpsService(){
+    public WpsService(CoreWorkspace coreWorkspace, DataManager dataManager){
+        this.dataManager = dataManager;
+        this.coreWorkspace = coreWorkspace;
         processManager = new ProcessManager();
+        dataProcessingManager = new DataProcessingManager(this);
+        if(!areScriptsCopied) {
+            setScriptFolder();
+        }
+        // Try to load previous state of the toolBox.
+        tbProperties = new Properties();
+        File propertiesFile = new File(coreWorkspace.getApplicationFolder() + File.separator + TOOLBOX_PROPERTIES);
+        if (propertiesFile.exists()) {
+            try {
+                tbProperties.load(new FileInputStream(propertiesFile));
+            } catch (IOException e) {
+                LoggerFactory.getLogger(ToolBox.class).warn("Unable to restore previous configuration of the ToolBox");
+                tbProperties = null;
+            }
+        }
+        if(tbProperties != null){
+            Object prop = tbProperties.getProperty(PROPERTY_SOURCES);
+            if(prop != null && !prop.toString().isEmpty()){
+                String str = prop.toString();
+                for(String s : str.split(";")){
+                    addLocalSource(URI.create(s), null, false);
+                }
+            }
+        }
+        //Find if the database used is H2 or not.
+        //If yes, make all the processes wait for the previous one.
+        try {
+            if(dataManager != null){
+                Connection connection = dataManager.getDataSource().getConnection();
+                isH2 = JDBCUtilities.isH2DataBase(connection.getMetaData());
+                if(isH2) {
+                    Statement statement = connection.createStatement();
+                    ResultSet result = statement.executeQuery("select VALUE from INFORMATION_SCHEMA.SETTINGS AS s where NAME = 'MVCC';");
+                    result.next();
+                    if (!result.getString(1).equals("TRUE")) {
+                        multiThreaded = false;
+                    }
+                    result = statement.executeQuery("select VALUE from INFORMATION_SCHEMA.SETTINGS AS s where NAME = 'MULTI_THREADED';");
+                    result.next();
+                    if (!result.getString(1).equals("1")) {
+                        multiThreaded = false;
+                    }
+                }
+            }
+        } catch (SQLException e) {
+            LoggerFactory.getLogger(ToolBox.class).error(e.getMessage());
+            multiThreaded = false;
+        }
+        if(!multiThreaded){
+            LoggerFactory.getLogger(ToolBox.class).warn("Warning, because of the H2 configuration," +
+                    " the toolbox won't be able to run more than one process at the same time.\n" +
+                    "Try to use the following setting for H2 : 'MVCC=TRUE; LOCK_TIMEOUT=100000; MULTI_THREADED=TRUE'");
+        }
+    }
+
+    public DataProcessingManager getDataProcessingManager() {
+        return dataProcessingManager;
+    }
+
+    public void dispose(){
+        //Try to save the local files loaded.
+        try {
+            if (tbProperties == null) {
+                tbProperties = new Properties();
+            }
+            tbProperties.setProperty(PROPERTY_SOURCES, processManager.getListSourcesAsString());
+            tbProperties.store(
+                    new FileOutputStream(coreWorkspace.getApplicationFolder() + File.separator + TOOLBOX_PROPERTIES),
+                    "Save of the OrbisGIS toolBox");
+        } catch (IOException e) {
+            LoggerFactory.getLogger(ToolBox.class).warn("Unable to save ToolBox state.");
+        }
+        areScriptsCopied = false;
+    }
+
+
+    @Reference
+    public void setCoreWorkspace(CoreWorkspace coreWorkspace) {
+        this.coreWorkspace = coreWorkspace;
+    }
+
+    public void unsetCoreWorkspace(CoreWorkspace coreWorkspace) {
+        this.coreWorkspace = null;
+    }
+
+    @Reference
+    public void setDriverFunctionContainer(DriverFunctionContainer driverFunctionContainer) {
+        this.driverFunctionContainer = driverFunctionContainer;
+    }
+
+    public void unsetDriverFunctionContainer(DriverFunctionContainer driverFunctionContainer) {
+        this.driverFunctionContainer = null;
+    }
+
+
+    /**
+     * Sets all the default OrbisGIS WPS script into the script folder of the .OrbisGIS folder.
+     */
+    private void setScriptFolder(){
+        //Sets the WPS script folder
+        File wpsScriptFolder = new File(coreWorkspace.getApplicationFolder(), WPS_SCRIPT_FOLDER);
+        //Empty the script folder or create it
+        if(wpsScriptFolder.exists()){
+            if(wpsScriptFolder.listFiles() != null) {
+                for (File f : wpsScriptFolder.listFiles()) {
+                    f.delete();
+                }
+            }
+        }
+        else{
+            if(!wpsScriptFolder.mkdir()){
+                LoggerFactory.getLogger(WpsService.class).warn("Unable to find or create a script folder.\n" +
+                        "No basic script will be available.");
+            }
+        }
+        if(wpsScriptFolder.exists() && wpsScriptFolder.isDirectory()){
+            try {
+                //Retrieve all the scripts url
+                String folderPath = ToolBox.class.getResource("scripts").getFile();
+                Enumeration<URL> enumUrl = FrameworkUtil.getBundle(WpsService.class).findEntries(folderPath, "*", false);
+                //For each url
+                while(enumUrl.hasMoreElements()){
+                    URL scriptUrl = enumUrl.nextElement();
+                    String scriptPath = scriptUrl.getFile();
+                    //Test if it's a groovy file
+                    if(scriptPath.endsWith("."+GROOVY_EXTENSION)){
+                        //If the script is already in the .OrbisGIS folder, remove it.
+                        for(File existingFile : wpsScriptFolder.listFiles()){
+                            if(existingFile.getName().endsWith(scriptPath) && existingFile.delete()){
+                                LoggerFactory.getLogger(WpsService.class).
+                                        warn("Replacing script "+existingFile.getName()+" by the default one");
+                            }
+                        }
+                        //Copy the script into the .OrbisGIS folder.
+                        OutputStream out = new FileOutputStream(
+                                new File(wpsScriptFolder.getAbsolutePath(),
+                                        new File(scriptPath).getName()));
+                        InputStream in = scriptUrl.openStream();
+                        IOUtils.copy(in, out);
+                        out.close();
+                        in.close();
+                    }
+                }
+            } catch (IOException e) {
+                LoggerFactory.getLogger(WpsService.class).warn("Unable to copy the scripts. \n" +
+                        "No basic script will be available. \n" +
+                        "Error : "+e.getMessage());
+            }
+        }
+        areScriptsCopied = true;
+        addLocalSource(wpsScriptFolder.toURI(), "orbisgis", true);
+    }
+    public void addLocalSource(URI uri, String iconName, boolean isDefaultScript){
+        File file = new File(uri);
+        if(file.isFile()){
+            return;
+        }
+        for(File f : file.listFiles()) {
+            addLocalScript(f, iconName, isDefaultScript);
+        }
+    }
+
+    public List<ProcessIdentifier> getProcessIdentifierFromParent(URI parent){
+        return processManager.getProcessIdentifierFromParent(parent);
+    }
+
+    public List<ProcessIdentifier> getAllProcessIdentifier(){
+        return processManager.getAllProcessIdentifier();
     }
 
     /**
@@ -56,5 +261,26 @@ public class WpsService {
 
     public void executeProcess(Process process, Map<URI, Object> dataMap, Map<String, Object> properties){
         processManager.executeProcess(process, dataMap, properties);
+    }
+
+    /**
+     * Save a geocatalog table into a file.
+     * @param uri URI where the table will be saved.
+     * @param tableName Name of the table to save.
+     */
+    public void saveURI(URI uri, String tableName){
+        try {
+            File f = new File(uri);
+            if(!f.exists()){
+                f.createNewFile();
+            }
+            //Find the good driver and save the file.
+            String extension = FilenameUtils.getExtension(f.getAbsolutePath());
+            DriverFunction driver = driverFunctionContainer.getImportDriverFromExt(
+                    extension, DriverFunction.IMPORT_DRIVER_TYPE.COPY);
+            driver.exportTable(dataManager.getDataSource().getConnection(), tableName, f, new EmptyProgressVisitor());
+        } catch (SQLException|IOException e) {
+            LoggerFactory.getLogger(ToolBox.class).error(e.getMessage());
+        }
     }
 }
