+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/execution/ExecutionWorker.java
@@ -83,10 +83,10 @@ public class ExecutionWorker extends SwingWorkerPM{
                     "Pre-processing");
             Map<URI, Object> stash = new HashMap<>();
             for(DescriptionType inputOrOutput : pee.getProcess().getOutput()){
-                stash.putAll(toolBox.getDataProcessingManager().preProcessData(inputOrOutput, dataMap));
+                stash.putAll(toolBox.getWpsService().getDataProcessingManager().preProcessData(inputOrOutput, dataMap));
             }
             for(DescriptionType inputOrOutput : pee.getProcess().getInput()){
-                stash.putAll(toolBox.getDataProcessingManager().preProcessData(inputOrOutput, dataMap));
+                stash.putAll(toolBox.getWpsService().getDataProcessingManager().preProcessData(inputOrOutput, dataMap));
             }
 
             //Execute the process and retrieve the groovy object.
@@ -100,10 +100,10 @@ public class ExecutionWorker extends SwingWorkerPM{
                     ProcessEditableElement.LogType.INFO,
                     "Post-processing");
             for(DescriptionType inputOrOutput : pee.getProcess().getOutput()){
-                toolBox.getDataProcessingManager().postProcessData(inputOrOutput, dataMap, stash);
+                toolBox.getWpsService().getDataProcessingManager().postProcessData(inputOrOutput, dataMap, stash);
             }
             for(DescriptionType inputOrOutput : pee.getProcess().getInput()){
-                toolBox.getDataProcessingManager().postProcessData(inputOrOutput, dataMap, stash);
+                toolBox.getWpsService().getDataProcessingManager().postProcessData(inputOrOutput, dataMap, stash);
             }
 
             //Print in the log the process execution end
