+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/view/ToolBox.java
@@ -76,11 +76,9 @@ public class ToolBox implements DockingPanel  {
     public static final String GROOVY_EXTENSION = "groovy";
     /** String reference of the ToolBox used for DockingFrame. */
     public static final String TOOLBOX_REFERENCE = "orbistoolbox";
-    private static final String WPS_SCRIPT_FOLDER = "Scripts";
     private static final String PROPERTY_SOURCES = "PROPERTY_SOURCES";
     private static final String TOOLBOX_PROPERTIES = "toolbox.properties";
     private static final String SETTINGS_TABLE = "SETTINGS";
-    private static boolean areScriptsCopied = false;
 
     /** Docking parameters used by DockingFrames. */
     private DockingPanelParameters parameters;
@@ -90,7 +88,6 @@ public class ToolBox implements DockingPanel  {
     private DataUIManager dataUIManager;
     /** Map containing the properties to apply for the Grovvy script execution. */
     private Map<String, Object> properties;
-    private DataProcessingManager dataProcessingManager;
     private CoreWorkspace coreWorkspace;
 
     /** EditableElement associated to the logEditor. */
@@ -108,18 +105,13 @@ public class ToolBox implements DockingPanel  {
     private static DataManager dataManager;
     /** OrbisGIS DriverFunctionContainer. */
     private static DriverFunctionContainer driverFunctionContainer;
-
-    /** ToolBox properties */
-    private Properties tbProperties;
-    /** List of process ended, waiting the 5 seconds before being removed.*/
-    boolean multiThreaded = true;
     /** True if the database is H2, false otherwise. */
     private boolean isH2;
     private WpsService wpsService;
 
     @Activate
     public void init(){
-        wpsService = new WpsService();
+        wpsService = new WpsService(coreWorkspace, dataManager);
         toolBoxPanel = new ToolBoxPanel(this);
         dataUIManager = new DataUIManager(this);
 
@@ -133,62 +125,18 @@ public class ToolBox implements DockingPanel  {
         parameters.setDockActions(dockingActions.getActions());
         dockingActions.addPropertyChangeListener(new ActionDockingListener(parameters));
 
-        dataProcessingManager = new DataProcessingManager(this);
         openEditorList = new ArrayList<>();
         lee = new LogEditableElement();
         le = null;
-
-        if(!areScriptsCopied) {
-            setScriptFolder();
-        }
-        // Try to load previous state of the toolBox.
-        tbProperties = new Properties();
-        File propertiesFile = new File(coreWorkspace.getApplicationFolder() + File.separator + TOOLBOX_PROPERTIES);
-        if (propertiesFile.exists()) {
-            try {
-                tbProperties.load(new FileInputStream(propertiesFile));
-            } catch (IOException e) {
-                LoggerFactory.getLogger(ToolBox.class).warn("Unable to restore previous configuration of the ToolBox");
-                tbProperties = null;
-            }
-        }
-        if(tbProperties != null){
-            Object prop = tbProperties.getProperty(PROPERTY_SOURCES);
-            if(prop != null && !prop.toString().isEmpty()){
-                String str = prop.toString();
-                for(String s : str.split(";")){
-                    addLocalSource(URI.create(s));
-                }
-            }
-        }
-        //Find if the database used is H2 or not.
-        //If yes, make all the processes wait for the previous one.
         try {
-            if(dataManager != null){
-                Connection connection = dataManager.getDataSource().getConnection();
-                isH2 = JDBCUtilities.isH2DataBase(connection.getMetaData());
-                if(isH2) {
-                    Statement statement = connection.createStatement();
-                    ResultSet result = statement.executeQuery("select VALUE from INFORMATION_SCHEMA.SETTINGS AS s where NAME = 'MVCC';");
-                    result.next();
-                    if (!result.getString(1).equals("TRUE")) {
-                        multiThreaded = false;
-                    }
-                    result = statement.executeQuery("select VALUE from INFORMATION_SCHEMA.SETTINGS AS s where NAME = 'MULTI_THREADED';");
-                    result.next();
-                    if (!result.getString(1).equals("1")) {
-                        multiThreaded = false;
-                    }
-                }
-            }
+            Connection connection = dataManager.getDataSource().getConnection();
+            isH2 = JDBCUtilities.isH2DataBase(connection.getMetaData());
         } catch (SQLException e) {
             LoggerFactory.getLogger(ToolBox.class).error(e.getMessage());
-            multiThreaded = false;
         }
-        if(!multiThreaded){
-            LoggerFactory.getLogger(ToolBox.class).warn("Warning, because of the H2 configuration," +
-                    " the toolbox won't be able to run more than one process at the same time.\n" +
-                    "Try to use the following setting for H2 : 'MVCC=TRUE; LOCK_TIMEOUT=100000; MULTI_THREADED=TRUE'");
+
+        for(ProcessIdentifier pi : wpsService.getAllProcessIdentifier()) {
+            toolBoxPanel.addLocalSource(pi);
         }
     }
 
@@ -196,86 +144,17 @@ public class ToolBox implements DockingPanel  {
         return wpsService;
     }
 
-    /**
-     * Sets all the default OrbisGIS WPS script into the script folder of the .OrbisGIS folder.
-     */
-    private void setScriptFolder(){
-        //Sets the WPS script folder
-        File wpsScriptFolder = new File(coreWorkspace.getApplicationFolder(), WPS_SCRIPT_FOLDER);
-        //Empty the script folder or create it
-        if(wpsScriptFolder.exists()){
-            if(wpsScriptFolder.listFiles() != null) {
-                for (File f : wpsScriptFolder.listFiles()) {
-                    f.delete();
-                }
-            }
-        }
-        else{
-            if(!wpsScriptFolder.mkdir()){
-                LoggerFactory.getLogger(ToolBox.class).warn("Unable to find or create a script folder.\n" +
-                        "No basic script will be available.");
-            }
-        }
-        if(wpsScriptFolder.exists() && wpsScriptFolder.isDirectory()){
-            try {
-                //Retrieve all the scripts url
-                String folderPath = ToolBox.class.getResource("scripts").getFile();
-                Enumeration<URL> enumUrl = FrameworkUtil.getBundle(ToolBox.class).findEntries(folderPath, "*", false);
-                //For each url
-                while(enumUrl.hasMoreElements()){
-                    URL scriptUrl = enumUrl.nextElement();
-                    String scriptPath = scriptUrl.getFile();
-                    //Test if it's a groovy file
-                    if(scriptPath.endsWith("."+GROOVY_EXTENSION)){
-                        //If the script is already in the .OrbisGIS folder, remove it.
-                        for(File existingFile : wpsScriptFolder.listFiles()){
-                            if(existingFile.getName().endsWith(scriptPath) && existingFile.delete()){
-                                LoggerFactory.getLogger(ToolBox.class).
-                                        warn("Replacing script "+existingFile.getName()+" by the default one");
-                            }
-                        }
-                        //Copy the script into the .OrbisGIS folder.
-                        OutputStream out = new FileOutputStream(
-                                new File(wpsScriptFolder.getAbsolutePath(),
-                                        new File(scriptPath).getName()));
-                        InputStream in = scriptUrl.openStream();
-                        IOUtils.copy(in, out);
-                        out.close();
-                        in.close();
-                    }
-                }
-            } catch (IOException e) {
-                LoggerFactory.getLogger(ToolBox.class).warn("Unable to copy the scripts. \n" +
-                        "No basic script will be available. \n" +
-                        "Error : "+e.getMessage());
-            }
-        }
-        areScriptsCopied = true;
-        addLocalSource(wpsScriptFolder.toURI(), "orbisgis", true);
-    }
+
 
     @Deactivate
-    public void dispose(){
+    public void dispose() {
         //Removes all the EditorDockable that were added
-        for(EditorDockable ed : openEditorList){
+        for (EditorDockable ed : openEditorList) {
             dockingManager.removeDockingPanel(ed.getDockingParameters().getName());
         }
         openEditorList = new ArrayList<>();
         toolBoxPanel.dispose();
-        areScriptsCopied = false;
-
-        //Try to save the local files loaded.
-        try {
-            if (tbProperties == null) {
-                tbProperties = new Properties();
-            }
-            tbProperties.setProperty(PROPERTY_SOURCES, toolBoxPanel.getListLocalSourcesAsString());
-            tbProperties.store(
-                    new FileOutputStream(coreWorkspace.getApplicationFolder() + File.separator + TOOLBOX_PROPERTIES),
-                    "Save of the OrbisGIS toolBox");
-        } catch (IOException e) {
-            LoggerFactory.getLogger(ToolBox.class).warn("Unable to save ToolBox state.");
-        }
+        wpsService.dispose();
     }
 
     @Override
@@ -460,14 +339,6 @@ public class ToolBox implements DockingPanel  {
         properties.remove("ds");
     }
 
-    @Reference
-    public void setCoreWorkspace(CoreWorkspace coreWorkspace) {
-        this.coreWorkspace = coreWorkspace;
-    }
-
-    public void unsetCoreWorkspace(CoreWorkspace coreWorkspace) {
-        this.coreWorkspace = null;
-    }
 
     @Reference
     public void setDataManager(DataManager dataManager) {
@@ -513,12 +384,17 @@ public class ToolBox implements DockingPanel  {
         ToolBox.driverFunctionContainer = null;
     }
 
-    public DriverFunctionContainer getDriverFunctionContainer(){
-        return driverFunctionContainer;
+    @Reference
+    public void setCoreWorkspace(CoreWorkspace coreWorkspace) {
+        this.coreWorkspace = coreWorkspace;
     }
 
-    public DataProcessingManager getDataProcessingManager() {
-        return dataProcessingManager;
+    public void unsetCoreWorkspace(CoreWorkspace coreWorkspace) {
+        this.coreWorkspace = null;
+    }
+
+    public DriverFunctionContainer getDriverFunctionContainer(){
+        return driverFunctionContainer;
     }
 
     public ToolBoxPanel getToolBoxPanel(){
@@ -683,27 +559,6 @@ public class ToolBox implements DockingPanel  {
             LoggerFactory.getLogger(ToolBox.class).error(e.getMessage());
         }
         return null;
-    }
-
-    /**
-     * Save a geocatalog table into a file.
-     * @param uri URI where the table will be saved.
-     * @param tableName Name of the table to save.
-     */
-    public void saveURI(URI uri, String tableName){
-        try {
-            File f = new File(uri);
-            if(!f.exists()){
-                f.createNewFile();
-            }
-            //Find the good driver and save the file.
-            String extension = FilenameUtils.getExtension(f.getAbsolutePath());
-            DriverFunction driver = driverFunctionContainer.getImportDriverFromExt(
-                    extension, DriverFunction.IMPORT_DRIVER_TYPE.COPY);
-            driver.exportTable(dataManager.getDataSource().getConnection(), tableName, f, new EmptyProgressVisitor());
-        } catch (SQLException|IOException e) {
-            LoggerFactory.getLogger(ToolBox.class).error(e.getMessage());
-        }
     }
 
     /**
